requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aenum==3.1.15
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==5.0.1
attrs==24.2.0
bitsandbytes==0.43.3
blendmodes==2022
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
clean-fid==0.1.35
click==8.1.7
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
cssselect2==0.7.0
cycler==0.12.1
Cython==3.0.11
deprecation==2.1.0
depth_anything @ https://github.com/huchenlei/Depth-Anything/releases/download/v1.0.0/depth_anything-2024.1.22.0-py2.py3-none-any.whl#sha256=26c1d38b8c3c306b4a2197d725a4b989ff65f7ebcf4fb5a96a1b6db7fbd56780
depth_anything_v2 @ https://github.com/MackinationsAi/UDAV2-ControlNet/releases/download/v1.0.0/depth_anything_v2-2024.7.1.0-py2.py3-none-any.whl#sha256=6848128867d1f7c7519d88df0f88bfab89100dc5225259c4d7cb90325c308c9f
diffusers==0.31.0
diskcache==5.6.3
easydict==1.13
einops==0.4.1
eval_type_backport==0.2.0
exceptiongroup==1.2.2
facexlib==0.3.0
fastapi==0.104.1
ffmpy==0.4.0
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.32
gradio==4.40.0
gradio_client==1.2.0
gradio_imageslider==0.0.20
gradio_rangeslider==0.0.6
h11==0.12.0
handrefinerportable @ https://github.com/huchenlei/HandRefinerPortable/releases/download/v1.0.1/handrefinerportable-2024.2.12.0-py2.py3-none-any.whl#sha256=1e6c702905919f4c49bcb2db7b20d334e8458a7555cd57630600584ec38ca6a9
httpcore==0.15.0
httpx==0.24.1
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
insightface @ https://github.com/Gourieff/Assets/raw/main/Insightface/insightface-0.7.3-cp310-cp310-win_amd64.whl#sha256=47aa0571b2aadd8545d4bc7615dfbc374c10180c283b7ac65058fcb41ed4df86
intel-openmp==2021.4.0
iopath==0.1.9
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
joblib==1.4.2
jsonmerge==1.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kornia==0.6.7
lark==1.1.2
lazy_loader==0.4
lightning-utilities==0.11.8
llvmlite==0.43.0
loadimg==0.1.2
lxml==5.3.0
manifold3d==2.5.1
mapbox_earcut==1.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.18
mkl==2021.4.0
ml_dtypes==0.5.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.2
omegaconf==2.2.3
onnx==1.17.0
onnxruntime==1.20.0
open-clip-torch==2.20.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.11
packaging==24.2
pandas==2.2.3
peft==0.13.2
piexif==1.1.3
Pillow==9.5.0
pillow-avif-plugin==1.4.3
portalocker==2.10.1
prettytable==3.12.0
propcache==0.2.0
protobuf==3.20.0
psutil==5.9.5
pycollada==0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytorch-lightning==1.9.4
pytz==2024.2
PyWavelets==1.7.0
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
resize-right==0.0.2
rich==13.9.4
rpds-py==0.21.0
Rtree==1.3.0
ruff==0.7.4
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
simsimd==6.0.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
spandrel==0.3.4
spandrel_extra_arches==0.1.1
starlette==0.27.0
stringzilla==3.10.10
svg.path==6.3
svglib==1.5.1
sympy==1.13.3
tabulate==0.9.0
tbb==2021.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.11
tinycss2==1.4.0
tokenizers==0.20.3
tomesd==0.1.3
tomlkit==0.12.0
torch==2.3.1+cu121
torchdiffeq==0.2.3
torchmetrics==1.6.0
torchsde==0.2.6
torchvision==0.18.1+cu121
tqdm==4.66.1
trampoline==0.1.2
transformers==4.46.1
trimesh==4.5.2
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
vhacdx==0.0.8.post1
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
xatlas==0.0.9
xxhash==3.5.0
yacs==0.1.8
yarl==1.17.1
zipp==3.21.0
